page abstraction improvements
